ゲスト側　完成

+++ janken/Scene/BaseScene.cpp
@@ -21,7 +21,7 @@ void BaseScene::DrawOwnScreen(void)
 
 }
 
-void BaseScene::Draw(float ex, float rad)
+void BaseScene::Draw(double ex, double rad)
 {
 	SetDrawScreen(DX_SCREEN_BACK);
 	DrawGraph(0, 0, screenID, true);
