+++ janken/Scene/GameScene.h
@@ -27,8 +27,8 @@ public:
 	void Init(void)override;
 	std::unique_ptr<BaseScene> Update(std::unique_ptr<BaseScene> own) override;
 	void Draw(void)override;
-	void Draw(float ex, float rad)override;
-	void SetBomb(int ownerID, int selfID, Vector2 pos,TimeP nowbool,float bombtime,int length,bool sendNet);
+	void Draw(double ex, double rad)override;
+	void SetBomb(int ownerID, int selfID, Vector2 pos,TimeP nowbool,double bombtime,int length,bool sendNet);
 	int BombCount(int ownerID);
 private:
 	std::vector<int> Image;
